% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_sync.R
\name{cli_sync_add_args}
\alias{cli_sync_add_args}
\title{Add 'sync' subparser arguments}
\usage{
cli_sync_add_args(subp, wf = NULL)
}
\arguments{
\item{subp}{Argparse subparsers object (from \verb{$add_subparsers()}).}

\item{wf}{(\code{character(1)} or \code{NULL})\cr
Workflow name. When non-NULL, \code{-w/--workflow} is not added.}
}
\description{
Registers arguments for the \code{sync} subcommand on an argparse subparsers
object. When \code{wf} is non-NULL the \code{-w/--workflow} argument is omitted
because the workflow is already fixed by the calling package.
}
\examples{
\dontrun{
p <- argparse::ArgumentParser(prog = "test.R", python_cmd = get_python())
subp <- p$add_subparsers(dest = "subparser_name")
cli_sync_add_args(subp, wf = "workflow1")
}
}
