% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_tidy.R
\name{cli_tidy_add_args}
\alias{cli_tidy_add_args}
\title{Add 'tidy' subparser arguments}
\usage{
cli_tidy_add_args(subp, wf = NULL)
}
\arguments{
\item{subp}{Argparse subparsers object (from \verb{$add_subparsers()}).}

\item{wf}{(\code{character(1)} or \code{NULL})\cr
Workflow name passed to \code{\link[=nemoverse_wf_dispatch]{nemoverse_wf_dispatch()}}.}
}
\description{
Registers arguments for the \code{tidy} subcommand on an argparse subparsers
object. When \code{wf} is non-NULL the \code{-w/--workflow} argument is omitted
because the workflow is already fixed by the calling package.
}
\examples{
\dontrun{
p <- argparse::ArgumentParser(prog = "test.R", python_cmd = get_python())
subp <- p$add_subparsers(dest = "subparser_name")
cli_tidy_add_args(subp, wf = "workflow1")
}
}
