% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_prep.R
\name{config_prep_multi}
\alias{config_prep_multi}
\title{Prepare config for multiple raw files}
\usage{
config_prep_multi(x)
}
\arguments{
\item{x}{(\code{tibble()})\cr
Tibble with columns \code{name}, \code{descr}, \code{pat}, \code{type}, and \code{path}.}
}
\value{
A list \code{list(tables = ...)} ready to write with \code{config_prep_write()}.
}
\description{
Prepare config for multiple raw files
}
\examples{
dir1 <-  "extdata/tool1/latest"
path1 <- system.file(dir1, "sampleA.tool1.table1.tsv", package = "nemo")
path2 <- system.file(dir1, "sampleA.tool1.table2.tsv", package = "nemo")
x <- tibble::tibble(
  name = c("table1", "table2"),
  descr = c("Table1 from Tool1.", "Table2 from Tool1."),
  pat = c("\\\\.tool1\\\\.table1\\\\.tsv$", "\\\\.tool1\\\\.table2\\\\.tsv$"),
  type = c("tsv", "tsv"),
  path = c(path1, path2)
)
config <- config_prep_multi(x)
}
