% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config.R
\name{config_sort_versions}
\alias{config_sort_versions}
\title{Sort version strings with "latest" always last}
\usage{
config_sort_versions(versions)
}
\arguments{
\item{versions}{(\code{character(n)})\cr
Vector of unique version strings (e.g. \code{c("v1.2.3", "latest")}) — typically from \code{unlist()}ing the \code{versions} list-col.}
}
\value{
Sorted character vector with \code{"latest"} last.
}
\description{
Sort version strings with "latest" always last
}
\keyword{internal}
