% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covarray.R
\name{covarray}
\alias{covarray}
\title{Format Covariates for NeMO}
\usage{
covarray(
  protocol = c("PCR_rep", "seq_read", "PCR_rep_seq_read"),
  array,
  cov_list = list(list(cov_data = NULL, level = "psi", dimension = "species"))
)
}
\arguments{
\item{protocol}{Character string. Specifies the modelling protocol to be used in \link{Nemodel}. \cr
Options are:
\itemize{
\item \code{'PCR_rep'}: Requires a 5D (N × I × J × K × C) presence/absence array
\item \code{'seq_read'}: Requires a 4D (N × I × J × C) sequence read count array
\item \code{'PCR_rep_seq_read'}: Requires a 5D (N × I × J × K × C) sequence read count array \cr
where:
\itemize{
\item N: Number of species
\item I: Number of sites
\item J: Number of samples
\item K: Number of PCR replicates
\item C: Number of campaigns
}
}}

\item{array}{Input data array to be used in \link{Nemodel}. Its required dimensionality depends on the selected \code{protocol} argument.}

\item{cov_list}{A list of sublists. Each sublist represent a covariate. Sublists must be attributed a unique name and contain the following components:
\itemize{
\item \code{cov_data}: Array of covariate values. These values should be either Boolean for categorical/semi-quantitative predictors or standardised values for quantitative predictors.
\item \code{level}: Character string. Specifies the hierarchical level where the covariate applies (\emph{e.g.}, \code{'psi'}, \code{'theta'}, \code{'p'}, \code{'phi'}).
\item \code{dimension}: Character string. Specifies the dimensions associated to the covariate. Acceptable values include:
\itemize{
\item Single dimensions: \code{'species'}, \code{'site'}, \code{'sample'}, \code{'replicate'}, or \code{'campaign'}.
\item Combined dimensions: Combinations of the above, except \code{'species'} (always stands alone). In combinations, terms must be separated by an underscore \code{'_'}, following this order: 1) \code{'site'} 2) \code{'sample'} 3) \code{'replicate'} 4) \code{'campaign'} (\emph{i.e.}, \code{'site_campaign'}).
}
}}
}
\value{
A structured class object with eight slots, organizing covariates across hierarchical levels for direct implementation in the NeMO modelling framework:
\itemize{
\item \code{psi_cov}: Covariates for spatial/methodological/temporal components of \eqn{\psi}.
\item \code{psi_cov_sp}: Species covariates of \eqn{\psi}.
\item \code{theta_cov}: Covariates for spatial/methodological/temporal components of \eqn{\theta}.
\item \code{theta_cov_sp}: Species covariates of \eqn{\theta}.
\item \code{p_cov}: Covariates for spatial/methodological/temporal components of \eqn{p}.
\item \code{p_cov_sp}: Species covariates of \eqn{p}.
\item \code{phi_cov}: Covariates for spatial/methodological/temporal components of \eqn{\varphi}.
\item \code{phi_cov_sp}: Species covariates of \eqn{\varphi}.
}
}
\description{
This function facilitates the integration of covariates into the
\strong{NeMO} modelling framework by formatting them into arrays compatible with \link{Nemodel}.
}
\details{
The \code{cov_data} array in each sublist must align with the specified \code{dimension}. For instance, if \code{dimension = 'site_sample_campaign'}, the covariate array should be structured such that dimension 1 corresponds to sites, dimension 2 corresponds to samples, and dimension 3 corresponds to campaigns.
}
\examples{
# Load input array
data(fish_PCR_rep)

# Load covariate data (Distance to sea)
data(distance_cov)

# Build the covariate array applied to sites on the psi level
covarray(protocol = 'PCR_rep',
         array = fish_PCR_rep,
         cov_list = list(Distance = list(cov_data = distance_cov$Distance,
                                         level = 'psi',
                                         dimension = 'site')))

}
\seealso{
\code{\link{Nemodel}}
}
