% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distance_cov}
\alias{distance_cov}
\title{Distance to Sea Covariate Data}
\format{
A dataframe with 10 rows and 1 column:
\describe{
\item{Distance}{Standardised distance to the sea for each site}
}
The \verb{$Distance} column can directly be used in \link{covarray}.
}
\usage{
data(distance_cov)
}
\description{
A dataframe containing the standardised distance to the sea for each of the 10 sites in the fish datasets.
}
\seealso{
\code{\link{covarray}}
}
\keyword{datasets}
