% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enframe_data}
\alias{enframe_data}
\title{Enframe Data}
\usage{
enframe_data(x)
}
\arguments{
\item{x}{(\code{list()})\cr
List to enframe.}
}
\value{
Enframed data with column name "data".
}
\description{
Enframe Data
}
