% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{file_hdr}
\alias{file_hdr}
\title{Get file header}
\usage{
file_hdr(fpath, delim = "\\t", n_max = 0, ...)
}
\arguments{
\item{fpath}{(\code{character(1)})\cr
File path.}

\item{delim}{(\code{character(1)})\cr
File delimiter.}

\item{n_max}{(\code{integer(1)})\cr
Maximum number of lines to read.}

\item{...}{Passed on to \code{readr::read_delim}.}
}
\description{
Returns the column names of a file without reading the entire file.
}
\examples{
dir1 <- system.file("extdata/tool1", package = "nemo")
fpath <- file.path(dir1, "latest", "sampleA.tool1.table1.tsv")
(hdr <- file_hdr(fpath))

}
