% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fish_PCR_rep}
\alias{fish_PCR_rep}
\title{Presence/Absence Data with Individually Indexed PCR Replicates}
\format{
A 5-dimensional array with:
\describe{
\item{Dimension 1}{Species (10)}
\item{Dimension 2}{Sites (10)}
\item{Dimension 3}{Samples (2)}
\item{Dimension 4}{PCR Replicates (5)}
\item{Dimension 5}{Campaigns (1)}
}
This dataset can directly be used in \link{Nemodel} with the \code{'PCR_rep'} protocol.
}
\usage{
data(fish_PCR_rep)
}
\description{
A 5D array containing presence/absence data for fish species across sites, samples, PCR replicates, and campaigns.
}
\seealso{
\code{\link{Nemodel}}
}
\keyword{datasets}
