% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fish_seq_read}
\alias{fish_seq_read}
\title{Sequence Read Count Data with Pooled PCR Replicates}
\format{
A 4-dimensional array with:
\describe{
\item{Dimension 1}{Species (10)}
\item{Dimension 2}{Sites (10)}
\item{Dimension 3}{Samples (2)}
\item{Dimension 4}{Campaigns (1)}
}
This dataset can directly be used in \link{Nemodel} with the \code{'seq_read'} protocol.
}
\usage{
data(fish_seq_read)
}
\description{
A 4D array containing sequence read count data for fish species across sites, samples, and campaigns.
}
\seealso{
\code{\link{Nemodel}}
}
\keyword{datasets}
