% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_resources.R
\name{min_resources}
\alias{min_resources}
\title{Calculate Minimum Resource Requirements for Reliable Species Detection}
\usage{
min_resources(model, resources = c("J"), conf = 0.95)
}
\arguments{
\item{model}{A fitted \link{Nemodel} object.}

\item{resources}{A character vector. Specifies the resource types to compute. Acceptable values include \code{'J'} (minimum samples), \code{'K'} (minimum PCR replicates), and \code{'M'} (minimum sequencing depth). Multiple resource types can be specified simultaneously (\emph{e.g.}, \code{resources = c('J', 'K', 'M')}).}

\item{conf}{Numeric. Confidence level, or probability to achieve to detect species when present (default: 0.95).}
}
\value{
A structured class object with three slots:
\itemize{
\item \code{J_min}: Minimum number of samples required to confidently detect species when present across sites.
\item \code{K_min}: Minimum number of replicates required to confidently detect species when present across sites and samples.
\item \code{M_min}: Minimum sequencing depth required to confidently detect species when present across sites, samples and replicates (except for \code{'seq_read'} protocol).
}

Each slot contains 3 arrays, corresponding to the median values and the lower and upper bounds of the 95\% highest density interval (HDI) of MCMC samples.
}
\description{
This function calculates the minimum resource requirements, \emph{i.e.}
samples \eqn{(J_{min})}, PCR replicates \eqn{(K_{min})}, and sequencing depth
\eqn{(M_{min})}, needed to confidently detect species when present with a specified
confidence level. This estimation relies on established probabilistic models,
enabling rigorous resource planning for ecological studies.
}
\details{
The function follows established equations from McArdle (1990) and probabilistic
models to compute:
\itemize{
\item \strong{Minimum number of samples} \eqn{\bold{(J_{min})}}: The function calculates
the minimum number of samples required to ensure that the probability of missing
species DNA in a sample is below 0.05 with 95\% confidence (can be computed for
any \code{protocol}):
\deqn{J_{min} \geq \frac{\ln(0.05)}{\ln(1 - \theta_{nijc})}}{J_min >= ln(0.05) / ln(1 - \theta_nijc)} \cr
where \eqn{\theta_{nijc}} is the probability of DNA collection for species \eqn{n} at site \eqn{i} in sample \eqn{j} during campaign \eqn{c}.
\cr
\item \strong{Minimum number of PCR replicates} \eqn{\bold{(K_{min})}}: The function calculates
the minimum number of PCR replicates required to ensure that the probability
of missing species DNA in a PCR replicate is below 0.05 with 95\% confidence (requires
a \link{Nemodel} object built with \code{'PCR_rep'} or \code{'PCR_rep_seq_read'} protocol):
\deqn{K_{min} \geq \frac{\ln(0.05)}{\ln(1 - p_{nijkc})}}{K_min >= ln(0.05) / ln(1 - p_nijkc)} \cr
where \eqn{p_{nijkc}} is the probability of DNA amplification for species \eqn{n} at site \eqn{i} in sample \eqn{j} in replicate \eqn{k} during campaign \eqn{c}.
\cr
\item \strong{Minimum sequencing depth} \eqn{\bold{(M_{min})}}: Following the multinomial
theorem, the function calculates the minimum sequencing depth required to ensure
that the probability of missing species DNA in a sample or in a PCR replicate
is below 0.05 (requires a \link{Nemodel} object built with \code{'seq_read'} or \code{'PCR_rep_seq_read'}
protocol):
\deqn{M_{min} \geq \frac{\ln(0.05)}{\ln(1 - \pi_{nijkc})}}{M_min >= ln(0.05) / ln(1 - \pi_nijkc)} \cr
where \eqn{\pi_{nijkc}} is the relative sequence read count for species \eqn{n} at site \eqn{i} in sample \eqn{j} in replicate \eqn{k} (except for \code{'seq_read'} protocol) during campaign \eqn{c}.
\cr
The function automatically performs these computations for precise and efficient
resource estimation.
}
}
\examples{
# Load fitted model
data(model_PCR_rep)

# Calculate the minimum number of samples and PCR replicates required to
# confidently detect species when present with 95\% confidence
min_resources(model = model_PCR_rep,
              resources = c('J', 'K'))

}
\seealso{
\code{Nemodel()}, \code{covarray()}
}
