% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_PCR_rep}
\alias{model_PCR_rep}
\title{Example model output using the \code{PCR_rep} protocol}
\format{
A structured class object (\link{Nemodel} object) with six slots:
\describe{
\item{model}{A fitted model in \code{rjags} format}
\item{protocol}{The modelling protocol used (\code{PCR_rep})}
\item{array}{The model input data array (\code{fish_PCR_rep})}
\item{covariates}{The covariate list implemented in the model (built from \code{distance_cov})}
\item{names}{A list containing species, site, sample, replicate, campaign and estimates' names}
\item{loglik}{\code{TRUE}: log-likelihood was stored for model comparison}
}
}
\usage{
data(model_PCR_rep)
}
\description{
This dataset contains the output from running \link{Nemodel} on \link{fish_PCR_rep} integrating \link{distance_cov} as covariate through \link{covarray}, using the \code{PCR_rep} protocol and the following parameters: \code{nb_iterations = 300}, \code{nb_burnin = 150}, \code{nb_thinning = 1}, \code{nb_chains = 2}.
}
\seealso{
\code{\link{Nemodel}}, \code{\link{covarray}}, \code{\link{fish_PCR_rep}}, \code{\link{distance_cov}}
}
\keyword{datasets}
