% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{nemo_assert_chr}
\alias{nemo_assert_chr}
\title{Assert character vector}
\usage{
nemo_assert_chr(x, arg = deparse(substitute(x)))
}
\arguments{
\item{x}{Object to check.}

\item{arg}{Name of \code{x} used in the error message.}
}
\value{
\code{x} invisibly on success.
}
\description{
Stops with an informative message if \code{x} is not a character vector.
}
