% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{nemo_assert_out_fmt}
\alias{nemo_assert_out_fmt}
\title{Assert output format}
\usage{
nemo_assert_out_fmt(format, choices = nemo_out_formats())
}
\arguments{
\item{format}{Output format to validate.}

\item{choices}{Available choices for valid output formats.}
}
\value{
\code{TRUE} invisibly on success.
}
\description{
Stops with an informative message if \code{format} is not a valid output format.
}
\examples{
nemo_assert_out_fmt("tsv")
}
