% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{nemo_assert_scalar_chr}
\alias{nemo_assert_scalar_chr}
\title{Assert scalar character}
\usage{
nemo_assert_scalar_chr(x, arg = deparse(substitute(x)))
}
\arguments{
\item{x}{Object to check.}

\item{arg}{Name of \code{x} used in the error message; defaults to the expression
passed as \code{x}.}
}
\value{
\code{x} invisibly on success.
}
\description{
Stops with an informative message if \code{x} is not a single non-NA character string.
}
\examples{
nemo_assert_scalar_chr("foo")
}
