% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemo_circle.R
\name{nemo_circle}
\alias{nemo_circle}
\title{Computes the largest empty circle which doesn't contain any points, inside a defined hull.}
\usage{
nemo_circle(points, hull, strict_inclusion, nmax_circles)
}
\arguments{
\item{points}{a set of points (an `sf` object).}

\item{hull}{external limits of the set of points (a polygon type `sf` object). Can be imported or computed with `nemo_hull` function, with the coordinate reference system as points object.}

\item{strict_inclusion}{TRUE if empty circle has to be entirely within the hull, FALSE otherwise.}

\item{nmax_circles}{number of empty circles in output.}
}
\value{
an `sf` object.
}
\description{

}
\examples{

data(points)
nemo_pts <-
nemo_circle(points = points \%>\% st_transform(2154),
           hull = hull_pts,
           strict_inclusion = T,
           nmax_circles = 1)



}
