% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{nemo_cli}
\alias{nemo_cli}
\title{Command line interface for nemo}
\usage{
nemo_cli(pkg = NULL, descr = NULL, wf = NULL)
}
\arguments{
\item{pkg}{(\code{character(1)})\cr
Package name.}

\item{descr}{(\code{character(1)})\cr
Description of the CLI program.}

\item{wf}{(\code{character(1)})\cr
Workflow name.}
}
\description{
Creates a command line interface (CLI) for the nemo package using the
argparse package.
}
