% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nemo_enframe}
\alias{nemo_enframe}
\title{Enframe Data}
\usage{
nemo_enframe(x)
}
\arguments{
\item{x}{(\code{list()})\cr
List to enframe.}
}
\value{
Enframed data with column name "data".
}
\description{
Enframe Data
}
