% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemo_hull.R
\name{nemo_hull}
\alias{nemo_hull}
\title{Computes the concave polygon for a set of points}
\usage{
nemo_hull(points, group_hull, concavity, length_threshold)
}
\arguments{
\item{points}{points for which the concave hull must be computed (an `sf` object).}

\item{concavity}{a relative measure of concavity. 1 results in a relatively detailed shape, Infinity results in a convex hull. You can use values lower than 1, but they can produce pretty crazy shapes.}

\item{length_threshold}{when a segment length is under this threshold, it stops being considered for further detalization. Higher values result in simpler shapes.}

\item{group}{optional : a field which splits the hull (a field of the `sf` object).}
}
\value{
an `sf` object.
}
\description{
Uses the `concaveman` function from [concaveman](https://github.com/mapbox/concaveman) and [concaveman in R](https://github.com/joelgombin/concaveman))
}
\examples{

data(points)
hull_pts <-
nemo_hull(points = points \%>\% st_transform(2154),
           #group_hull = k,
           concavity =2,
           length_threshold = 10)



}
