% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{nemo_osfx}
\alias{nemo_osfx}
\title{Construct Output File Paths with Format Suffix}
\usage{
nemo_osfx(fpfix, format)
}
\arguments{
\item{fpfix}{(\code{character(n)})\cr
Vector of one or more file prefixes e.g. /path/to/foo}

\item{format}{(\code{character(1)})\cr
Output format. One of tsv, csv, parquet, rds, or db.}
}
\value{
Character vector of output file paths
}
\description{
Construct Output File Paths with Format Suffix
}
\examples{
fpfix <- "path/to/foo"
format <- "tsv"
(o <- nemo_osfx(fpfix, format))
}
