% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{nemo_stop}
\alias{nemo_stop}
\title{Stop with call suppressed}
\usage{
nemo_stop(...)
}
\arguments{
\item{...}{Passed to \code{\link[=stop]{stop()}}.}
}
\value{
Does not return.
}
\description{
Thin wrapper around \code{\link[=stop]{stop()}} with \code{call. = FALSE}.
}
