% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nemodel.R
\name{Nemodel}
\alias{Nemodel}
\title{Bayesian Multispecies Occupancy Model for eDNA Data.}
\usage{
Nemodel(
  protocol = c("PCR_rep", "seq_read", "PCR_rep_seq_read"),
  array,
  covariates = NULL,
  name = "model",
  nb_iterations = 1000,
  nb_burnin = floor(nb_iterations/2),
  nb_thinning = max(1, floor((nb_iterations - nb_burnin)/1000)),
  nb_chains = 2,
  parallel = FALSE,
  loglik = FALSE,
  latent = NULL,
  posterior = NULL,
  tau = 1,
  delta = 1,
  ...
)
}
\arguments{
\item{protocol}{Character string. Specifies the modeling protocol. \cr
Options are:
\itemize{
\item \code{'PCR_rep'}: Requires a 5D (N × I × J × K × C) presence/absence input array
\item \code{'seq_read'}: Requires a 4D (N × I × J × C) sequence read count input array
\item \code{'PCR_rep_seq_read'}: Requires a 5D (N × I × J × K × C) sequence read count input array \cr
where:
\itemize{
\item N: Number of species.
\item I: Number of sites.
\item J: Number of samples.
\item K: Number of PCR replicates.
\item C: Number of campaigns.
}
}}

\item{array}{Input data array. Its required dimensionality depends on the selected \code{protocol} argument.}

\item{covariates}{Optional. A preprocessed covariate list generated using \link{covarray}.}

\item{name}{Character string specifying the filename and path for recording the BUGS language model file.}

\item{nb_iterations}{Integer. Total number of MCMC iterations to run (default: 2000).}

\item{nb_burnin}{Integer. Number of initial burn-in iterations (default: \code{floor(nb_iterations / 2)}).}

\item{nb_thinning}{Integer. Thinning interval to reduce autocorrelation in MCMC samples (default: \code{max(1, floor((nb_iterations - nb_burnin) / 1000))}).}

\item{nb_chains}{Integer. Number of independent MCMC chains to compute (default: 3).}

\item{parallel}{Logical. If \code{TRUE}, enables parallel computation to speed up sampling for large datasets (default: \code{FALSE}).}

\item{loglik}{Logical. If \code{TRUE}, computes and saves log-likelihoods for model comparison using \link{WAIC} (default: \code{FALSE}).}

\item{latent}{Optional character vector. Specifies latent arrays (\emph{e.g.}, \code{'Z'}, \code{'A'}, \code{'W'}, \code{'S'}, \code{'Y'}) to save.}

\item{posterior}{Optional character vector. Specifies posterior distributions of key parameters (\emph{e.g.}, \code{'psi'}, \code{'theta'}, \code{'p'}, \code{'pi'}, \code{'phi'}) to save.}

\item{tau}{Numeric. Precision parameter for normal (hyper)priors (default: 1).}

\item{delta}{Numeric. Dispersion parameter of the negative binomial distribution \eqn{S} (only for \code{'seq_read'} or \code{'PCR_rep_seq_read'} protocols, default: 1).}

\item{...}{Additional arguments (related to \link{jags} or \link{jags.parallel} functions from \strong{R2jags}).}
}
\value{
A structured class object (\link{Nemodel} object) with five slots:
\itemize{
\item \code{model}: The fitted model output in \code{rjags} format, with a BUGS language text file stored at the specified location (\code{name} argument).
\item \code{protocol}: The modelling protocol used (\emph{i.e.}, \code{'PCR_rep'}, \code{'seq_read'}, or \code{'PCR_rep_seq_read'}).
\item \code{array}: The model input data array.
\item \code{covariates}: The covariate list implemented in the model.
\item \code{names}: A list containing species, site, sample, replicate, and campaign names. This slot also contains:
\itemize{
\item \code{estimates_sp}: Names associated with the species-level random effects.
\item \code{estimates}: Names associated with the intercepts and other random effects.
}
\item \code{loglik}: Logical. If \code{TRUE}, log-likelihood values are stored for use in model comparison.
\item \code{delta}: Numeric. Dispersion parameter used to model the negative binomial distribution \eqn{S}.
}
}
\description{
This function is the core of the \strong{NeMO} package, designed for
Bayesian modelling of multispecies occupancy in eDNA metabarcoding studies. It
provides flexibility to fit different modelling protocols, accommodating various
study designs and data structures. The function uses the \strong{R2jags} package for
MCMC sampling.
}
\details{
This function provides flexibility for different study designs by allowing users to model occupancy using either presence/absence arrays (\code{PCR_rep}) or sequence read counts (\code{seq_read}, \code{PCR_rep_seq_read}).
Covariates can be incorporated to account for external predictors.
The MCMC sampling process can be customized using multiple control parameters (\code{nb_iterations}, \code{nb_burnin}, \code{nb_thinning}, etc.), and parallel computation can be enabled with \code{parallel = TRUE}.
The function also allows users to store log-likelihoods, latent arrays, and posterior distributions for downstream analysis.
}
\examples{
# Load input array
data(fish_PCR_rep)

# Load covariate data (Distance to sea)
data(distance_cov)

# Build the covariate array applied to sites on the psi level
covariates <- covarray(protocol = 'PCR_rep',
                       array = fish_PCR_rep,
                       cov_list = list(Distance = list(cov_data = distance_cov$Distance,
                                                       level = 'psi',
                                                       dimension = 'site')))

# Run the 'PCR_rep' model on the input array with the distance covariate
Nemodel(protocol = 'PCR_rep',
        array = fish_PCR_rep,
        covariates = covariates)
unlink('model.txt')

}
\seealso{
\code{\link{WAIC}}, \code{\link{covarray}}
}
