% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEMO.R
\docType{data}
\name{NUM.NEIGHBORS.RATIO}
\alias{NUM.NEIGHBORS.RATIO}
\title{Num neighbors ratio}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
NUM.NEIGHBORS.RATIO
}
\description{
Num neighbors ratio
}
\keyword{datasets}
