% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_file}
\alias{parse_file}
\title{Parse file}
\usage{
parse_file(fpath, pname, schemas_all, delim = "\\t", ...)
}
\arguments{
\item{fpath}{(\code{character(1)})\cr
File path.}

\item{pname}{(\code{character(1)})\cr
Parser name (e.g. "breakends" - see docs).}

\item{schemas_all}{(\code{tibble()})\cr
Tibble with name, version and schema list-col.}

\item{delim}{(\code{character(1)})\cr
File delimiter.}

\item{...}{Passed on to \code{readr::read_delim}.}
}
\description{
Parses files.
}
\examples{
path <- system.file("extdata/tool1", package = "nemo")
x <- Tool$new("tool1", pkg = "nemo", path)
schemas_all <- x$raw_schemas_all
f <- function(ver, tbl) file.path(path, ver, paste0("sampleA.tool1.", tbl, ".tsv"))
# table1: three versions with different column sets
(d1_v123 <- parse_file(f("v1.2.3", "table1"), "table1", schemas_all))
(d1_v456 <- parse_file(f("v4.5.6", "table1"), "table1", schemas_all))
(d1_lat  <- parse_file(f("latest", "table1"), "table1", schemas_all))
# table2: two versions (v1.0.0 drops metricB)
(d2_v1  <- parse_file(f("v1.0.0", "table2"), "table2", schemas_all))
(d2_lat <- parse_file(f("latest", "table2"), "table2", schemas_all))

}
