% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_file_keyvalue}
\alias{parse_file_keyvalue}
\title{Parse Key-Value file}
\usage{
parse_file_keyvalue(fpath, pname, schemas_all, delim = "\\t", ...)
}
\arguments{
\item{fpath}{(\code{character(1)})\cr
File path.}

\item{pname}{(\code{character(1)})\cr
Parser name.}

\item{schemas_all}{(\code{tibble()})\cr
Tibble with name, version and schema list-col.}

\item{delim}{(\code{character(1)})\cr
File delimiter.}

\item{...}{Passed on to \code{readr::read_delim}.}
}
\description{
Parses files with no header and two columns representing key-value pairs,
pivoted wide and typed per the matching raw schema.
}
\examples{
path <- system.file("extdata/tool1", package = "nemo")
x <- Tool1$new(path)
schemas_all <- x$config$get_schemas_raw()
pname <- "table3"
f <- function(ver) file.path(path, ver, "sampleA.tool1.table3.tsv")
# v1.0.0: 3 key-value pairs (SampleID, QCStatus, TotalReads)
(d3_v1  <- parse_file_keyvalue(f("v1.0.0"), pname, schemas_all))
# latest: 5 key-value pairs
(d3_lat <- parse_file_keyvalue(f("latest"), pname, schemas_all))

}
