% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_file_nohead}
\alias{parse_file_nohead}
\title{Parse headless file}
\usage{
parse_file_nohead(fpath, schema, delim = "\\t", ...)
}
\arguments{
\item{fpath}{(\code{character(1)})\cr
File path.}

\item{schema}{(\code{tibble()})\cr
Schema tibble with version and schema list-col.}

\item{delim}{(\code{character(1)})\cr
File delimiter.}

\item{...}{Passed on to \code{readr::read_delim}.}
}
\description{
Parses files with no column names.
}
\examples{
path <- system.file("extdata/tool1", package = "nemo")
x <- Tool$new("tool1", pkg = "nemo", path)
schemas_all <- x$config$schemas_raw
pname <- "table4"
fpath_latest <- file.path(path, "latest", "sampleA.tool1.table4.tsv")
fpath_v1 <- file.path(path, "v1.0.0", "sampleA.tool1.table4.tsv")
schema_latest <- schemas_all |>
  dplyr::filter(.data$name == pname, .data$version == "latest") |>
  dplyr::select("version", "schema")
schema_v1 <- schemas_all |>
  dplyr::filter(.data$name == pname, .data$version == "v1.0.0") |>
  dplyr::select("version", "schema")
(d_latest <- parse_file_nohead(fpath_latest, schema_latest))
(d_v1 <- parse_file_nohead(fpath_v1, schema_v1))

}
