% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkg_found}
\alias{pkg_found}
\title{Check if Package is Installed}
\usage{
pkg_found(p)
}
\arguments{
\item{p}{(\code{character(1)})\cr
Package name.}
}
\value{
\code{TRUE} if the package is installed, \code{FALSE} otherwise.
}
\description{
Check if an R package is installed.
}
\examples{
pkg_found("base")
pkg_found("somefakepackagename")
}
