% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_s3.R
\name{s3_file_presignedurl}
\alias{s3_file_presignedurl}
\title{S3 Generate Presigned URL}
\usage{
s3_file_presignedurl(client, s3path, expiry_seconds = 604800)
}
\arguments{
\item{client}{S3 client. Make sure you use \code{signature_version = "s3v4"} (see example).}

\item{s3path}{Full path to S3 object.}

\item{expiry_seconds}{Number of seconds the presigned URL is valid for (3600 = 1 hour).}
}
\value{
An S3 presigned URL.
}
\description{
S3 Generate Presigned URL
}
\examples{
\dontrun{
client <- paws.storage::s3(paws.storage::config(signature_version = "s3v4"))
s3path <- "s3://bucket1/path/to/file.tsv"
s3_file_presignedurl(client, s3path)
}

}
