% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{schema_guess}
\alias{schema_guess}
\title{Guess Schema}
\usage{
schema_guess(pname, cnames, schemas_all)
}
\arguments{
\item{pname}{(\code{character(1)})\cr
Parser name.}

\item{cnames}{(\code{character(n)})\cr
Column names.}

\item{schemas_all}{(\code{tibble()})\cr
Tibble with name, version and schema list-col.}
}
\description{
Given a tibble of available schemas, filters to the one
matching the given column names. Errors out if unsuccessful.
}
\examples{
dir1 <- system.file("extdata/tool1", package = "nemo")
fpath1 <- file.path(dir1, "latest", "sampleA.tool1.table1.tsv")
fpath2 <- file.path(dir1, "v1.2.3", "sampleA.tool1.table1.tsv")
pname <- "table1"
cnames1 <- file_hdr(fpath1)
cnames2 <- file_hdr(fpath2)
conf <- Config$new("tool1", pkg = "nemo")
schemas_all <- conf$get_raw_schemas_all()
(s1 <- schema_guess(pname, cnames1, schemas_all))
(s2 <- schema_guess(pname, cnames2, schemas_all))

}
