% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_vis.R
\name{schema_version_colour}
\alias{schema_version_colour}
\title{Version colour}
\usage{
schema_version_colour(v)
}
\arguments{
\item{v}{version string.}
}
\value{
A CSS colour string.
}
\description{
Version colour
}
\keyword{internal}
