% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_vis.R
\name{schema_version_pill}
\alias{schema_version_pill}
\title{Version pill: \code{latest} green, others grey}
\usage{
schema_version_pill(v)
}
\arguments{
\item{v}{version string.}
}
\value{
An htmltools span.
}
\description{
Version pill: \code{latest} green, others grey
}
\keyword{internal}
