% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_tbl_version_attr}
\alias{set_tbl_version_attr}
\title{Set Table Version Attribute}
\usage{
set_tbl_version_attr(tbl, v, x = "file_version")
}
\arguments{
\item{tbl}{(\code{tibble()})\cr
Table with a version attribute.}

\item{v}{(\code{character(1)})\cr
Version string to set.}

\item{x}{(\code{character(1)})\cr
Name of the attribute to retrieve.}
}
\description{
Set the version attribute from a table.
}
\examples{
d <- tibble::tibble(a = 1:3, b = letters[1:3])
v <- "v1.2.3"
d <- set_tbl_version_attr(d, v)
(a <- attr(d, "file_version"))

}
