% 
% TODO: Manual for class StaticNEM 
% 
\name{StaticNEMDesign}
\alias{StaticNEMDesign}
\title{
	StaticNEMDesign class
}
\description{
	This is an class defining the design information for the systematic perturbation experiments.  
}
\usage{
	StaticNEMDesign(design=NULL)
}
\arguments{
	\item{design}{
		the design table including sample name, group name and replicate id
	}
}
\value{
	a StaticNEMDesign object
}
\details{
	This class includes a few operations for design tables.
}
\author{
	Xin Wang
}
\examples{
	# demo("static_nem", "nemo")
}

