% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config.R
\name{config_prep_raw}
\alias{config_prep_raw}
\title{Prepare config from raw file}
\usage{
config_prep_raw(path, name, descr, pat, type = "tsv", v = "latest", ...)
}
\arguments{
\item{path}{(\code{character(1)})\cr
File path.}

\item{name}{(\code{character(1)})\cr
File nickname.}

\item{descr}{(\code{character(1)})\cr
File description.}

\item{pat}{(\code{character(1)})\cr
File pattern.}

\item{type}{(\code{character(1)})\cr
File type.}

\item{v}{(\code{character(1)})\cr
File version.}

\item{...}{Passed on to \code{readr::read_delim}.}
}
\value{
A named list with the config info.
}
\description{
Prepares config from raw file.
}
\examples{
path <- system.file("extdata", "tool1/latest/sampleA.tool1.table1.tsv", package = "nemo")
name <- "table1"
descr <- "Table1 from Tool1."
pat <- "\\\\.tool1\\\\.table1\\\\.tsv$"
l <- config_prep_raw(path, name, descr, pat)
}
