% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config.R
\name{config_prep_write}
\alias{config_prep_write}
\title{Write config to YAML file}
\usage{
config_prep_write(x, out)
}
\arguments{
\item{x}{(\code{list()})\cr
Config list, generated by \code{config_prep_multi()}.}

\item{out}{(\code{character(1)})\cr
Output file path.}
}
\value{
Nothing, called for side effects.
}
\description{
Write config to YAML file
}
\examples{
dir1 <- "extdata/tool1/latest"
path1 <- system.file(dir1, "sampleA.tool1.table1.tsv", package = "nemo")
path2 <- system.file(dir1, "sampleA.tool1.table2.tsv", package = "nemo")
x <- tibble::tibble(
  name = c("table1", "table2"),
  descr = c("Table1 from Tool1.", "Table2 from Tool1."),
  pat = c("\\\\.tool1\\\\.table1\\\\.tsv$", "\\\\.tool1\\\\.table2\\\\.tsv$"),
  type = c("txt", "txt"),
  path = c(path1, path2)
)
config <- config_prep_multi(x)
out <- tempfile(fileext = ".yaml")
config_prep_write(config, out)
}
