% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config.R
\name{config_prep_write}
\alias{config_prep_write}
\title{Write config to YAML file}
\usage{
config_prep_write(x, out)
}
\arguments{
\item{x}{(\code{list()})\cr
Config list, generated by \code{config_prep_multi()}.}

\item{out}{(\code{character(1)})\cr
Output file path.}
}
\value{
Nothing, called for side effects.
}
\description{
Write config to YAML file
}
