% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tbl_version_attr}
\alias{get_tbl_version_attr}
\title{Get Table Version Attribute}
\usage{
get_tbl_version_attr(tbl, x = "file_version")
}
\arguments{
\item{tbl}{(\code{tibble()})\cr
Table with a version attribute.}

\item{x}{(\code{character(1)})\cr
Name of the attribute to retrieve.}
}
\description{
Get the version attribute from a table.
}
\examples{
path <- system.file("extdata/tool1", package = "nemo")
path2 <- file.path(path, "v1.2.3", "sampleA.tool1.table1.tsv")
x <- Tool1$new(path)$tidy(keep_raw = TRUE)
ind <- which(x$tbls$path == path2)
stopifnot(length(ind) == 1)
(v <- get_tbl_version_attr(x$tbls$raw[[ind]]))

}
