\name{NeMO-package}
\alias{NeMO}
\docType{package}
\title{Nested eDNA Metabarcoding Occupancy}
\description{
    This package provides a flexible Bayesian framework for modeling multispecies
    site occupancy from eDNA metabarcoding data. It supports diverse study designs,
    including individually indexed or pooled PCR replicates, enabling inference from
    either presence/absence or sequence read count data. It allows covariate implementation
    and subsequent model comparison using the Watanabe-Akaike Information Criterion.
    Additionally, it makes possible the estimate of the minimum resources required
    to confidently detect species.
}
\author{
    Bastien Macé \email{bastien.mace.49@gmail.com}
}
