% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidywigits-package.R
\docType{package}
\name{nemo-package}
\alias{nemo}
\alias{nemo-package}
\title{nemo: Tidy and explore outputs from bioinformatic pipelines}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides the base components for tidying and exploring outputs from bioinformatic pipelines.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/umccr/nemo}
  \item \url{https://umccr.github.io/nemo/}
  \item Report bugs at \url{https://github.com/umccr/nemo/issues}
}

}
\author{
\strong{Maintainer}: Peter Diakumis \email{peterdiakumis@gmail.com} (\href{https://orcid.org/0000-0002-7502-7545}{ORCID})

}
\keyword{internal}
