% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{nemo_log}
\alias{nemo_log}
\title{Log a message with a specified level}
\usage{
nemo_log(level, msg, ...)
}
\arguments{
\item{level}{The log level ("DEBUG", "INFO", "WARN", "ERROR", "FATAL").}

\item{msg}{The message to log. Use \link{sprintf} formatting.}

\item{...}{Values to format into the message. See \link{sprintf} for details.}
}
\description{
Log a message with a specified level
}
