% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{nemo_metadata}
\alias{nemo_metadata}
\title{Title}
\usage{
nemo_metadata(files, pkgs, input_id, output_id, input_dir, output_dir)
}
\arguments{
\item{files}{(\code{tibble()})\cr
Written files.}

\item{pkgs}{(\code{character(n)})\cr
Packages to include versions of.}

\item{input_id}{Input ID to use for the dataset (e.g. \code{run123}).}

\item{output_id}{(\code{character(1)})\cr
Output ID to use for the dataset (e.g. \code{run123}).}

\item{input_dir}{(\code{character(n)})\cr
Input directory (can be multiple).}

\item{output_dir}{(\code{character(1)})\cr
Output directory.}
}
\value{
List of metadata.
}
\description{
Title
}
\examples{
files <- tibble::tibble(
  tbl_name = c("purple_qc", "amber_qc"),
  prefix = c("S123", "S123"),
  outpath = c("S123_purple_qc.tsv", "S123_amber_qc.tsv")
)
pkgs <- c("nemo", "tidywigits")
input_id <- "run123"
output_id <- ulid::ulid()
input_dir <- "/path/to/wigits/run123"
output_dir <- "/path/to/nemo/outputs/run123"
nemo_metadata(files, pkgs, input_id, output_id, input_dir, output_dir)
}
