% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{nemo_write}
\alias{nemo_write}
\title{Write data}
\usage{
nemo_write(d, fpfix = NULL, format = "tsv", dbconn = NULL, dbtab = NULL)
}
\arguments{
\item{d}{(\code{data.frame()})\cr
A data.frame (or tibble) with tidy data.}

\item{fpfix}{(\code{character(1)})\cr
File prefix. The file extension is generated automatically via the \code{format}
argument. For a format of db, this is inserted into the \code{nemo_pfix} column.}

\item{format}{(\code{character(1)})\cr
Output format. One of tsv, csv, parquet, rds, or db.}

\item{dbconn}{(\code{DBIConnection(1)})\cr
Database connection object (see \code{DBI::dbConnect}). Used only when format is db.}

\item{dbtab}{(\code{character(1)})\cr
Database table name (see \code{DBI::dbWriteTable}). Used only when format is db.}
}
\description{
Writes tabular data in the given format.
}
\examples{
d <- tibble::tibble(name = "foo", data = 123)
fpfix <- file.path(tempdir(), "data_test1")
format <- "csv"
nemo_write(d = d, fpfix = fpfix, format = format)
(res <- readr::read_csv(glue::glue("{fpfix}.csv.gz"), show_col_types = FALSE))
\dontshow{if (RPostgres::postgresHasDefault()) withAutoprint(\{ # examplesIf}
# for database writing
con <- DBI::dbConnect(RPostgres::Postgres())
tbl_nm <- "awesome_tbl"
nemo_write(d = d, fpfix = basename(fpfix), format = "db", dbconn = con, dbtab = tbl_nm)
DBI::dbListTables(con)
DBI::dbReadTable(con, tbl_nm)
DBI::dbDisconnect(con)
\dontshow{\}) # examplesIf}
}
