% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nemoverse_wf_dispatch}
\alias{nemoverse_wf_dispatch}
\title{Nemoverse Workflow Dispatcher}
\usage{
nemoverse_wf_dispatch(wf = NULL)
}
\arguments{
\item{wf}{Workflow name.}
}
\value{
The nemo workflow class to initiate.
}
\description{
Dispatches the nemoverse workflow class based on the chosen workflow.
}
\examples{
wf <- "basemean"
(fun <- nemoverse_wf_dispatch(wf))
}
