% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_file}
\alias{parse_file}
\title{Parse file}
\usage{
parse_file(fpath, pname, schemas_all, delim = "\\t", ...)
}
\arguments{
\item{fpath}{(\code{character(1)})\cr
File path.}

\item{pname}{(\code{character(1)})\cr
Parser name (e.g. "breakends" - see docs).}

\item{schemas_all}{(\code{tibble()})\cr
Tibble with name, version and schema list-col.}

\item{delim}{(\code{character(1)})\cr
File delimiter.}

\item{...}{Passed on to \code{readr::read_delim}.}
}
\description{
Parses files.
}
\examples{
path <- system.file("extdata/tool1", package = "nemo")
x <- Tool$new("tool1", pkg = "nemo", path)
schemas_all <- x$raw_schemas_all
pname <- "table1"
fpath <- file.path(path, "latest", "sampleA.tool1.table1.tsv")
(d <- parse_file(fpath, pname, schemas_all))

}
