% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_file_keyvalue}
\alias{parse_file_keyvalue}
\title{Parse Key-Value file}
\usage{
parse_file_keyvalue(fpath, pname, schemas_all, delim = "\\t", ...)
}
\arguments{
\item{fpath}{(\code{character(1)})\cr
File path.}

\item{pname}{(\code{character(1)})\cr
Parser name.}

\item{schemas_all}{(\code{tibble()})\cr
Tibble with name, version and schema list-col.}

\item{delim}{(\code{character(1)})\cr
File delimiter.}

\item{...}{Passed on to \code{readr::read_delim}.}
}
\description{
Parses files with no header and two columns representing key-value pairs.
}
\examples{
dir1 <- system.file("extdata/tool1", package = "nemo")
fpath <- file.path(dir1, "latest", "sampleA.tool1.table3.tsv")
x <- Tool1$new(dir1)
schemas_all <- x$raw_schemas_all
pname <- "table3"
(d <- parse_file_keyvalue(fpath, pname, schemas_all))

}
