% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_file_nohead}
\alias{parse_file_nohead}
\title{Parse headless file}
\usage{
parse_file_nohead(fpath, schema, delim = "\\t", ...)
}
\arguments{
\item{fpath}{(\code{character(1)})\cr
File path.}

\item{schema}{(\code{tibble()})\cr
Schema tibble with version and schema list-col.}

\item{delim}{(\code{character(1)})\cr
File delimiter.}

\item{...}{Passed on to \code{readr::read_delim}.}
}
\description{
Parses files with no column names.
}
