% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_file_nohead}
\alias{parse_file_nohead}
\title{Parse headless file}
\usage{
parse_file_nohead(fpath, schema, delim = "\\t", ...)
}
\arguments{
\item{fpath}{(\code{character(1)})\cr
File path.}

\item{schema}{(\code{tibble()})\cr
Schema tibble with version and schema list-col.}

\item{delim}{(\code{character(1)})\cr
File delimiter.}

\item{...}{Passed on to \code{readr::read_delim}.}
}
\description{
Parses files with no column names.
}
\examples{
path <- system.file("extdata/tool1", package = "nemo")
x <- Tool$new("tool1", pkg = "nemo", path = path)
schema <- x$raw_schemas_all |>
       dplyr::filter(.data$name == "table3") |>
       dplyr::select("version", "schema")
fpath <- file.path(path, "latest", "sampleA.tool1.table3.tsv")
(d <- parse_file_nohead(fpath, schema))

}
