% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tool1.R
\name{Tool1}
\alias{Tool1}
\title{Tool1 Object}
\description{
Parses and tidies output files from Tool1. Table schemas and file patterns
are defined in \code{inst/config/tools/tool1/schema.yaml}. Custom parse and tidy
methods are provided for tables that require non-standard handling.
}
\examples{
indir <- system.file("extdata/tool1", package = "nemo")
dir1 <- tempdir()
obj1 <- Tool1$new(indir)

p3 <- system.file("extdata/tool1/latest/sampleA.tool1.table3.tsv", package = "nemo")
(tidy3 <- obj1$tidy_table3(p3))

obj1$run(output_dir = dir1, format = "parquet", input_id = "run1")
(lf <- list.files(dir1, pattern = "tool1.*parquet", full.names = FALSE))

obj2 <- Tool1$new(indir)$tidy()
}
\section{Super class}{
\code{\link[nemo:Tool]{nemo::Tool}} -> \code{Tool1}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Tool1-new}{\code{Tool1$new()}}
\item \href{#method-Tool1-tidy_table3}{\code{Tool1$tidy_table3()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="nemo" data-topic="Tool" data-id="filter_files"><a href='../../nemo/html/Tool.html#method-Tool-filter_files'><code>nemo::Tool$filter_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nemo" data-topic="Tool" data-id="get_metadata"><a href='../../nemo/html/Tool.html#method-Tool-get_metadata'><code>nemo::Tool$get_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nemo" data-topic="Tool" data-id="get_tbls"><a href='../../nemo/html/Tool.html#method-Tool-get_tbls'><code>nemo::Tool$get_tbls()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nemo" data-topic="Tool" data-id="list_files"><a href='../../nemo/html/Tool.html#method-Tool-list_files'><code>nemo::Tool$list_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nemo" data-topic="Tool" data-id="print"><a href='../../nemo/html/Tool.html#method-Tool-print'><code>nemo::Tool$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nemo" data-topic="Tool" data-id="run"><a href='../../nemo/html/Tool.html#method-Tool-run'><code>nemo::Tool$run()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nemo" data-topic="Tool" data-id="tidy"><a href='../../nemo/html/Tool.html#method-Tool-tidy'><code>nemo::Tool$tidy()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nemo" data-topic="Tool" data-id="write"><a href='../../nemo/html/Tool.html#method-Tool-write'><code>nemo::Tool$write()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tool1-new"></a>}}
\if{latex}{\out{\hypertarget{method-Tool1-new}{}}}
\subsection{Method \code{new()}}{
Create a new Tool1 object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tool1$new(path = NULL, files_tbl = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{(\code{character(1)})\cr
Output directory of tool. If \code{files_tbl} is supplied, this is ignored.}

\item{\code{files_tbl}}{(\code{tibble(n)})\cr
Tibble of files from \code{\link[=list_files_dir]{list_files_dir()}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{R6::R6Class()})\cr
R6 object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tool1-tidy_table3"></a>}}
\if{latex}{\out{\hypertarget{method-Tool1-tidy_table3}{}}}
\subsection{Method \code{tidy_table3()}}{
Tidy \code{table3.tsv} file with type conversion enabled.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tool1$tidy_table3(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(\code{character(1)} or \code{tibble()})\cr
Path to file or already parsed tibble.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{tibble()})\cr
Tidy data in enframed tibble.
}
}
}
