% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tool1.R
\name{Tool1}
\alias{Tool1}
\title{Tool1 Object}
\description{
Tool1 file parsing and manipulation.
}
\examples{
cls <- Tool1
indir <- system.file("extdata/tool1", package = "nemo")
odir <- tempdir()
id <- "tool1_run1"
obj <- cls$new(indir)
obj$nemofy(odir = odir, format = "parquet", id = id)
list.files(odir, pattern = "parquet", full.names = FALSE)
}
\section{Super class}{
\code{\link[nemo:Tool]{nemo::Tool}} -> \code{Tool1}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Tool1-new}{\code{Tool1$new()}}
\item \href{#method-Tool1-parse_table1}{\code{Tool1$parse_table1()}}
\item \href{#method-Tool1-tidy_table1}{\code{Tool1$tidy_table1()}}
\item \href{#method-Tool1-parse_table2}{\code{Tool1$parse_table2()}}
\item \href{#method-Tool1-tidy_table2}{\code{Tool1$tidy_table2()}}
\item \href{#method-Tool1-clone}{\code{Tool1$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="nemo" data-topic="Tool" data-id=".eval_func"><a href='../../nemo/html/Tool.html#method-Tool-.eval_func'><code>nemo::Tool$.eval_func()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nemo" data-topic="Tool" data-id=".parse_file"><a href='../../nemo/html/Tool.html#method-Tool-.parse_file'><code>nemo::Tool$.parse_file()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nemo" data-topic="Tool" data-id=".parse_file_nohead"><a href='../../nemo/html/Tool.html#method-Tool-.parse_file_nohead'><code>nemo::Tool$.parse_file_nohead()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nemo" data-topic="Tool" data-id=".tidy_file"><a href='../../nemo/html/Tool.html#method-Tool-.tidy_file'><code>nemo::Tool$.tidy_file()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nemo" data-topic="Tool" data-id="filter_files"><a href='../../nemo/html/Tool.html#method-Tool-filter_files'><code>nemo::Tool$filter_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nemo" data-topic="Tool" data-id="list_files"><a href='../../nemo/html/Tool.html#method-Tool-list_files'><code>nemo::Tool$list_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nemo" data-topic="Tool" data-id="nemofy"><a href='../../nemo/html/Tool.html#method-Tool-nemofy'><code>nemo::Tool$nemofy()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nemo" data-topic="Tool" data-id="print"><a href='../../nemo/html/Tool.html#method-Tool-print'><code>nemo::Tool$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nemo" data-topic="Tool" data-id="tidy"><a href='../../nemo/html/Tool.html#method-Tool-tidy'><code>nemo::Tool$tidy()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nemo" data-topic="Tool" data-id="write"><a href='../../nemo/html/Tool.html#method-Tool-write'><code>nemo::Tool$write()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tool1-new"></a>}}
\if{latex}{\out{\hypertarget{method-Tool1-new}{}}}
\subsection{Method \code{new()}}{
Create a new Tool1 object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tool1$new(path = NULL, files_tbl = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{(\code{character(1)})\cr
Output directory of tool. If \code{files_tbl} is supplied, this basically gets
ignored.}

\item{\code{files_tbl}}{(\code{tibble(n)})\cr
Tibble of files from \code{\link[=list_files_dir]{list_files_dir()}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tool1-parse_table1"></a>}}
\if{latex}{\out{\hypertarget{method-Tool1-parse_table1}{}}}
\subsection{Method \code{parse_table1()}}{
Read \code{table1.tsv} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tool1$parse_table1(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(\code{character(1)})\cr
Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tool1-tidy_table1"></a>}}
\if{latex}{\out{\hypertarget{method-Tool1-tidy_table1}{}}}
\subsection{Method \code{tidy_table1()}}{
Tidy \code{table1.tsv} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tool1$tidy_table1(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(\code{character(1)})\cr
Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tool1-parse_table2"></a>}}
\if{latex}{\out{\hypertarget{method-Tool1-parse_table2}{}}}
\subsection{Method \code{parse_table2()}}{
Read \code{table2.tsv} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tool1$parse_table2(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(\code{character(1)})\cr
Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tool1-tidy_table2"></a>}}
\if{latex}{\out{\hypertarget{method-Tool1-tidy_table2}{}}}
\subsection{Method \code{tidy_table2()}}{
Tidy \code{table2.tsv} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tool1$tidy_table2(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(\code{character(1)})\cr
Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tool1-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Tool1-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tool1$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
