% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workflow1.R
\name{Workflow1}
\alias{Workflow1}
\title{Workflow1 Object}
\description{
Workflow1 file parsing and manipulation.
}
\examples{
path <- system.file("extdata/tool1", package = "nemo")
odir <- tempdir()
id <- "workflow1_run1"
w <- Workflow1$new(path)
x <- w$nemofy(diro = odir, format = "parquet", input_id = id)
(lf <- list.files(odir, pattern = "tool1.*parquet", full.names = FALSE))
}
\section{Super class}{
\code{\link[nemo:Workflow]{nemo::Workflow}} -> \code{Workflow1}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Workflow1-new}{\code{Workflow1$new()}}
\item \href{#method-Workflow1-clone}{\code{Workflow1$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="nemo" data-topic="Workflow" data-id="filter_files"><a href='../../nemo/html/Workflow.html#method-Workflow-filter_files'><code>nemo::Workflow$filter_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nemo" data-topic="Workflow" data-id="get_metadata"><a href='../../nemo/html/Workflow.html#method-Workflow-get_metadata'><code>nemo::Workflow$get_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nemo" data-topic="Workflow" data-id="get_raw_schemas_all"><a href='../../nemo/html/Workflow.html#method-Workflow-get_raw_schemas_all'><code>nemo::Workflow$get_raw_schemas_all()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nemo" data-topic="Workflow" data-id="get_tbls"><a href='../../nemo/html/Workflow.html#method-Workflow-get_tbls'><code>nemo::Workflow$get_tbls()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nemo" data-topic="Workflow" data-id="get_tidy_schemas_all"><a href='../../nemo/html/Workflow.html#method-Workflow-get_tidy_schemas_all'><code>nemo::Workflow$get_tidy_schemas_all()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nemo" data-topic="Workflow" data-id="list_files"><a href='../../nemo/html/Workflow.html#method-Workflow-list_files'><code>nemo::Workflow$list_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nemo" data-topic="Workflow" data-id="nemofy"><a href='../../nemo/html/Workflow.html#method-Workflow-nemofy'><code>nemo::Workflow$nemofy()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nemo" data-topic="Workflow" data-id="print"><a href='../../nemo/html/Workflow.html#method-Workflow-print'><code>nemo::Workflow$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nemo" data-topic="Workflow" data-id="tidy"><a href='../../nemo/html/Workflow.html#method-Workflow-tidy'><code>nemo::Workflow$tidy()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nemo" data-topic="Workflow" data-id="write"><a href='../../nemo/html/Workflow.html#method-Workflow-write'><code>nemo::Workflow$write()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Workflow1-new"></a>}}
\if{latex}{\out{\hypertarget{method-Workflow1-new}{}}}
\subsection{Method \code{new()}}{
Create a new Workflow1 object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Workflow1$new(path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{(\code{character(n)})\cr
Path(s) to Workflow1 results.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Workflow1-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Workflow1-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Workflow1$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
